day7
Given an array, of size , reverse it.

Example: If array, , after reversing it, the array should be, .

Input Format

The first line contains an integer, , denoting the size of the array. The next line contains  space-separated integers denoting the elements of the array.

Constraints


, where  is the  element of the array.

Output Format

The output is handled by the code given in the editor, which would print the array.